requirements.txt
Django==5.0.1
pymysql==1.1.0
Pillow
python-decouple==3.8
django-ckeditor==6.7.0
django-crispy-forms==2.1
crispy-tailwind==1.0.3
pandas
openpyxl
